+++ src/main/java/com/hydata/intelligence/platform/repositories/UserRepository.java
@@ -28,7 +28,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
 	
 	@QueryHints(value = {@QueryHint(name = HINT_COMMENT ,value= "a query for pageable")})
 	@Query("select u from User u where u.name like concat('%' ,?1,'%') and u.isvalid = ?2")
-	Page<User> findByNameAndIsvalid(String name,Integer isValid,Pageable page);
+	Page<User> findByNameAndIsvalid(String name,Byte isValid,Pageable page);
 	
 	@Query("select u from User u where u.name = ?1 and u.pwd = ?2")
 	Optional<User> findByNameAndPWD(String name,String pwd);
